Conqa Test
Users can:

Add a truck with a number, record the truck region, the total ice cream
in each vat(5x), and the number of unopened sprinkles.

Users can edit the existing trucks data and load test data from a .js
file.

Truck data is persistent and uses firebase as a backend service.

TODO - Add form validation for input values.
TO: fix the progress bars to display how fill each vat in the truck is.